+++ src/compiler/parse_stmt.c
@@ -1421,7 +1421,6 @@ Ast *parse_stmt(ParseContext *c)
 		case TOKEN_CT_STRINGIFY:
 		case TOKEN_CT_TERNARY:
 		case TOKEN_CT_VAARG:
-		case TOKEN_CT_VACOUNT:
 		case TOKEN_FALSE:
 		case TOKEN_INTEGER:
 		case TOKEN_LENGTHOF:
@@ -1463,7 +1462,6 @@ Ast *parse_stmt(ParseContext *c)
 		case TOKEN_CT_ENDSWITCH:
 		case TOKEN_CT_EXEC:
 		case TOKEN_CT_INCLUDE:
-		case TOKEN_CT_VASPLAT:
 		case TOKEN_DIV:
 		case TOKEN_DIV_ASSIGN:
 		case TOKEN_DOCS_END: